Make SVG files usable with CSS
SVG files are now located beneath the _includes folder. This allows
them to be inserted as raw HTML rather than as images, which lets
them be modifiable by CSS. Now, they can change colors when hovered,
grow in size, shrink, whatever.

Additionally, this is exemplified by the ISU logo in the footer. It
changes color to yellow when hovered.